+++ pkg/constant.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 const (
 	ConfigMapLabel string = "nacos.io/owned-by-dc"
