+++ pkg/nacos/auth/auth_provider.go
