+++ pkg/nacos/object_reference_wrapper.go
@@ -3,7 +3,7 @@ package nacos
 import (
 	"context"
 	"fmt"
-	"github.com/nacos-group/nacos-controller/internal"
+	"github.com/nacos-group/nacos-controller/pkg"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -216,8 +216,8 @@ func (cmw *ConfigMapWrapper) ensureOwnerLabel() error {
 	if cmw.cm.Labels == nil {
 		cmw.cm.Labels = map[string]string{}
 	}
-	if v, ok := cmw.cm.Labels[internal.ConfigMapLabel]; !ok || v != cmw.owner.GetName() {
-		cmw.cm.Labels[internal.ConfigMapLabel] = cmw.owner.GetName()
+	if v, ok := cmw.cm.Labels[pkg.ConfigMapLabel]; !ok || v != cmw.owner.GetName() {
+		cmw.cm.Labels[pkg.ConfigMapLabel] = cmw.owner.GetName()
 		return cmw.Flush()
 	}
 	return nil
